Fixed a constraint issue with the pokemon_collection table and converted the class components to functional components

+++ backend/src/main/java/com/pokemon_team_builder/backend/model/repository/PokemonRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PokemonRepository extends JpaRepository<Pokemon, Integer> {
+public interface PokemonRepository extends JpaRepository<Pokemon, Long> {
     Optional<Pokemon> findPokemonByName(String name);
 }
